requirements.txt
slack-cleaner2
pyyaml
